requirements.txt
Pillow>=9.0.0
numpy>=1.23.0
opencv-python>=4.7.0

